fixes folder layout
