+++ app/components/header/Header.tsx
@@ -1,0 +1,21 @@
+const Header = () => (
+  <ul className="NavigationBar">
+    <li>
+      <a href="#boxe">Boxe</a>
+    </li>
+    <li>|</li>
+    <li>
+      <a href="#muaythai">Muay Thai</a>
+    </li>
+    <li>|</li>
+    <li>
+      <a href="#jiujitsu">Jiu Jitsu</a>
+    </li>
+    <li>|</li>  
+    <li>
+      <a href="#contato">Contato</a>
+    </li>
+  </ul>
+);
+
+export default Header;
